Refactor: add services folder, move trust logic into it
